backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-dotenv==1.0.0
